Move entity browser classes to their own file.
